Move Hasher into abstractions as it's got no dependencies, is a format utility
