+++ NuKeeper/Engine/BranchNamer.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using NuKeeper.Abstractions.Formats;
 using NuKeeper.Inspection.RepositoryInspection;
 
 namespace NuKeeper.Engine
